+++ tests/testthat/test-global_map.R
@@ -3,13 +3,7 @@ context("global_map")
 
 data <- suppressWarnings(
   EstZoonoticTB::link_data(verbose = FALSE) %>% 
-  dplyr::filter(!is.na(tb_z_prop)) %>% 
-  dplyr::mutate(country = country %>% 
-                  as.character) %>% 
-  dplyr::group_by(country) %>% 
-  dplyr::arrange(desc(year)) %>% 
-  dplyr::slice(1) %>% 
-  dplyr::ungroup()
+  EstZoonoticTB::get_latest_data(tb_z_prop)
 )